pyproject.toml
[tool.poetry]
name = "red-gpt5assistant"
version = "0.1.0"
description = "Red DiscordBot cog: GPT-5 Assistant using OpenAI Responses API with native tools"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "gpt5assistant" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
red-discordbot = ">=3.5.0"
openai = ">=1.99.0"
httpx = ">=0.27.0"
anyio = ">=4.0.0"
cachetools = ">=5.3.0"
tenacity = ">=8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2.0"
pytest-asyncio = ">=0.23.0"
pytest-cov = ">=5.0.0"
mypy = ">=1.10.0"
ruff = ">=0.5.0"
types-cachetools = ">=5.3.0.7"
types-requests = "*"

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B","SIM","TID","C4"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

